refactor(coin-zcash): consolidate routing on zcashShielded (LIVE-34556)
- route Zcash accounts to coin-zcash through the existing zcashShielded flag and
  drop the zcashCoinModule flag; the flag state is mirrored in coin-zcash's own
  constants, which is what bridge/impl.ts reads
- host a single Zcash engine: coin-zcash owns the zcash: IPC prefix and the
  zcash-utility bundle, and coin-bitcoin's Zcash adapter is no longer built
- tolerate a coin-zcash transaction in the shared bitcoin send UI: with no utxo
  strategy, hide the advanced fee modes and show the estimated ZIP-317 fee
- restore the coin-bitcoin behaviours the port had dropped: full viewing key and
  estimation recipient hooks, pre-broadcast double-spend guard, single device
  session for xpub derivation, NU5 activation date, transparent
  transactionSequenceNumber, upfront send capability check
- keep logic/ to what CoinModuleApi reaches and move the rest to bridge/; the
  double-spend guard now runs inside logic/transaction/broadcast.ts and the
  balance arithmetic is shared from logic/account/getBalance.ts
- implement getBalance and listOperations over the Ledger explorer address
  routes (transparent scope, cursor propagated); shielded value stays out of
  reach there, as notes are encrypted to a viewing key the signature has no room
  for
- reject a t-address whose checksum does not hold, which the ZIP-316 prefix
  classifier was letting through
