+++ apps/ledger-live-desktop/tools/rspack/rspack.renderer.ts
@@ -83,9 +83,15 @@ export function createRendererConfig(
         ...commonConfig.resolve?.alias,
         LLD: path.resolve(lldRoot, "src", "mvvm"),
         "styled-components": styledComponentsPath,
-        // Route `ZCash` to the IPC client in the renderer so the
-        // `zcash-utils` .node addon stays out of the bundle (see
-        // `@ledgerhq/coin-bitcoin/chain-adapters/zcash/ipc/main-host`).
+        // Route `ZCash` to the IPC client in the renderer so the `zcash-utils`
+        // .node addon stays out of the bundle: it is hosted in a UtilityProcess,
+        // reached over the `zcash:*` channels the main process registers (see
+        // `@ledgerhq/coin-zcash/network/ipc/main-host`).
+        "@ledgerhq/coin-zcash/network/ZCash$": "@ledgerhq/coin-zcash/network/ZCashIPC",
+        // The Zcash chain-adapter of `coin-bitcoin` needs the same alias -- its
+        // lazy `zcash-native` chunk is still resolved at build time, addon
+        // included -- even though its shielded path is no longer served: the
+        // `zcashShielded` flag routes those accounts to coin-zcash.
         "@ledgerhq/coin-bitcoin/chain-adapters/zcash/ZCash$":
           "@ledgerhq/coin-bitcoin/chain-adapters/zcash/ZCashIPC",
         // Fix tests/time.js import for TIMEMACHINE feature
