+++ libs/coin-modules/coin-zcash/src/bridge/estimateMaxSpendable.ts
@@ -1,7 +1,7 @@
 import { getMainAccount } from "@ledgerhq/ledger-wallet-framework/account/index";
 import type { AccountBridge } from "@ledgerhq/types-live";
 import type { Transaction, ZcashAccount } from "../types/bridge";
-import { collectSpendableNotes } from "../logic/operations";
+import { collectSpendableNotes } from "./operations";
 import {
   estimateMaxSpendableAmount,
   estimateMaxSpendableTransparent,
