+++ libs/coin-modules/coin-zcash/src/bridge/transaction-details.ts
