--- libs/coin-modules/coin-zcash/src/logic/utils.ts
@@ -1,6 +1,0 @@
-import { isValidZcashAddress } from "./validateAddress";
-
-/** Synchronous recipient-validity check used by the bridge (getTransactionStatus). */
-export function isRecipientValid(recipient: string): boolean {
-  return isValidZcashAddress(recipient);
-}
