+++ libs/coin-modules/coin-zcash/src/network/ipc/contract.ts
@@ -47,7 +47,16 @@ export type RequestId = string;
 
 // --- Renderer <-> Main (ipcRenderer.invoke / webContents.send) ---
 
-/** Channel names -- centralized to keep renderer/main/utility in lock-step. */
+/**
+ * Channel names -- centralized to keep renderer/main/utility in lock-step.
+ *
+ * The `zcash:` prefix is shared with the identical contract the Zcash
+ * chain-adapter of `coin-bitcoin` declares, and that is deliberate: only one
+ * host may `ipcMain.handle` a given channel (the second registration throws),
+ * and the host the desktop app registers is this module's. The adapter's own
+ * host is no longer built or registered -- it was only ever reached from its
+ * shielded path, which the `zcashShielded` feature flag now routes here.
+ */
 export const ZCASH_IPC = {
   /** invoke -> number (chain tip height) */
   getChainTip: "zcash:getChainTip",
