--- shared/feature-flags/src/flags/team-coin-integration/zcashCoinModule.ts
@@ -1,2 +1,0 @@
-import { flag } from "../../define";
-export const zcashCoinModule = flag();
